+++ packages/suite/src/actions/wallet/sendFormActions.ts
@@ -1,8 +1,12 @@
 import BigNumber from 'bignumber.js';
 import { SEND } from '@wallet-actions/constants';
 import { getOutput } from '@wallet-utils/sendFormUtils';
-import { formatNetworkAmount, getFiatValue, getNetworkAmount } from '@wallet-utils/accountUtils';
-import { getAccountKey } from '@suite-utils/reducerUtils';
+import {
+    formatNetworkAmount,
+    getFiatValue,
+    getNetworkAmount,
+    getAccountKey,
+} from '@wallet-utils/accountUtils';
 import { Output, InitialState, FeeLevel } from '@wallet-types/sendForm';
 import { Account } from '@wallet-types';
 import { Dispatch, GetState } from '@suite-types';
