+++ packages/suite/src/utils/wallet/__tests__/accountUtils.test.ts
@@ -1,10 +1,11 @@
 import * as accountUtils from '../accountUtils';
 import * as fixtures from './fixtures/accountUtils';
+import { NETWORKS } from '@wallet-config';
 import { Account } from '@wallet-types';
 
 const { intlMock } = global.JestMocks;
 
-describe('accountUtils', () => {
+describe('account utils', () => {
     fixtures.parseBIP44Path.forEach(f => {
         it('accountUtils.parseBIP44Path', () => {
             expect(accountUtils.parseBIP44Path(f.path)).toEqual(f.result);
@@ -92,5 +93,156 @@ describe('accountUtils', () => {
                 state: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
             }),
         );
+    });
+
+    fixtures.getAccountTransactions.forEach(f => {
+        it(`getAccountTransactions${f.testName}`, () => {
+            expect(
+                accountUtils.getAccountTransactions(
+                    // @ts-ignore TODO: Missing isAddress on TransactionTarget coming from connect/blockbook?
+                    f.transactions as TransactionsState['transactions'],
+                    f.account as Account,
+                ),
+            ).toEqual(f.result);
+        });
+    });
+
+    it('getSelectedAccount null', () => {
+        const res = accountUtils.getSelectedAccount([], undefined, undefined);
+        expect(res).toBeNull();
+    });
+
+    it('getSelectedNetwork', () => {
+        const res = accountUtils.getSelectedNetwork(NETWORKS, 'btc');
+        if (res) {
+            expect(res.name).toEqual('Bitcoin');
+        } else {
+            expect(res).toBeNull();
+        }
+        expect(accountUtils.getSelectedNetwork(NETWORKS, 'doesntexist')).toBeNull();
+    });
+
+    it('getAccountHash', () => {
+        expect(accountUtils.getAccountKey('descriptor', 'symbol', 'deviceState')).toEqual(
+            'descriptor-symbol-deviceState',
+        );
+    });
+
+    fixtures.enhanceTransaction.forEach(f => {
+        it('enhanceTransaction', () => {
+            // @ts-ignore
+            expect(accountUtils.enhanceTransaction(f.tx, f.account)).toEqual(f.result);
+        });
+    });
+
+    it('getSelectedAccount', () => {
+        expect(
+            accountUtils.getSelectedAccount(
+                [
+                    global.JestMocks.getWalletAccount({
+                        descriptor:
+                            'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
+                        symbol: 'btc',
+                        index: 0,
+                    }),
+                    global.JestMocks.getWalletAccount({
+                        symbol: 'btc',
+                        descriptor: '123',
+                        accountType: 'normal',
+                        index: 1,
+                    }),
+                ],
+                global.JestMocks.getSuiteDevice({
+                    state: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
+                }),
+                {
+                    symbol: 'btc',
+                    accountIndex: 1,
+                    accountType: 'normal',
+                },
+            ),
+        ).toEqual(
+            global.JestMocks.getWalletAccount({
+                symbol: 'btc',
+                descriptor: '123',
+                accountType: 'normal',
+                index: 1,
+            }),
+        );
+
+        expect(
+            accountUtils.getSelectedAccount(
+                [
+                    global.JestMocks.getWalletAccount({
+                        descriptor:
+                            'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
+                        symbol: 'btc',
+                        index: 0,
+                    }),
+                    global.JestMocks.getWalletAccount({
+                        symbol: 'btc',
+                        descriptor: '123',
+                        accountType: 'normal',
+                        index: 1,
+                    }),
+                ],
+                global.JestMocks.getSuiteDevice({
+                    state: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
+                }),
+                undefined,
+            ),
+        ).toBeNull();
+
+        expect(
+            accountUtils.getSelectedAccount(
+                [
+                    global.JestMocks.getWalletAccount({
+                        descriptor:
+                            'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
+                        symbol: 'btc',
+                        index: 0,
+                    }),
+                    global.JestMocks.getWalletAccount({
+                        symbol: 'btc',
+                        descriptor: '123',
+                        accountType: 'normal',
+                        index: 1,
+                    }),
+                ],
+                undefined,
+                {
+                    symbol: 'btc',
+                    accountIndex: 1,
+                    accountType: 'normal',
+                },
+            ),
+        ).toBeNull();
+
+        expect(
+            accountUtils.getSelectedAccount(
+                [
+                    global.JestMocks.getWalletAccount({
+                        descriptor:
+                            'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
+                        symbol: 'btc',
+                        index: 0,
+                    }),
+                    global.JestMocks.getWalletAccount({
+                        symbol: 'btc',
+                        descriptor: '123',
+                        accountType: 'normal',
+                        index: 1,
+                    }),
+                ],
+                global.JestMocks.getSuiteDevice({
+                    state: '7dcccffe70d8bb8bb28a2185daac8e05639490eee913b326097ae1d73abc8b4f',
+                }),
+                {
+                    symbol: 'btc',
+                    accountIndex: 3,
+                    accountType: 'normal',
+                },
+            ),
+        ).toBeNull();
     });
 });
