+++ packages/suite/src/utils/wallet/accountUtils.ts
@@ -1,3 +1,8 @@
+import {
+    State as TransactionsState,
+    WalletAccountTransaction,
+} from '@wallet-reducers/transactionReducer';
+import { AccountTransaction } from 'trezor-connect';
 import l10nMessages from '@wallet-views/account/messages';
 import { InjectedIntl } from 'react-intl';
 import BigNumber from 'bignumber.js';
@@ -216,4 +221,94 @@ export const isAddressInAccount = (
 export const getAccountDevice = (devices: AppState['devices'], account: Account) => {
     const device = devices.find(d => d.state === account.deviceState);
     return device;
+};
+
+export const getSelectedAccount = (
+    accounts: Account[],
+    device: AppState['suite']['device'],
+    routerParams: AppState['router']['params'],
+) => {
+    if (!device || !routerParams) return null;
+
+    // TODO: imported accounts
+    // imported account index has 'i' prefix
+    // const isImported = /^i\d+$/i.test(routerParams.accountIndex);
+    // const index: number = isImported
+    //     ? parseInt(routerParams.accountIndex.substr(1), 10)
+    //     : parseInt(routerParams.accountIndex, 10);
+
+    return (
+        accounts.find(
+            a =>
+                a.index === routerParams.accountIndex &&
+                a.symbol === routerParams.symbol &&
+                a.accountType === routerParams.accountType &&
+                a.deviceState === device.state,
+        ) || null
+    );
+};
+
+export const getSelectedNetwork = (networks: Network[], symbol: string) => {
+    return networks.find(c => c.symbol === symbol) || null;
+};
+
+/**
+ * Returns a string used as an index to separate txs for given account inside a transactions reducer
+ *
+ * @param {string} descriptor
+ * @param {string} symbol
+ * @param {string} deviceState
+ * @returns {string}
+ */
+export const getAccountKey = (descriptor: string, symbol: string, deviceState: string) => {
+    return `${descriptor}-${symbol}-${deviceState}`;
+};
+
+export const getAccountTransactions = (
+    transactions: TransactionsState['transactions'],
+    account: Account,
+) => {
+    const accountHash = getAccountKey(account.descriptor, account.symbol, account.deviceState);
+    return transactions[accountHash] || [];
+};
+
+/**
+ * Formats amounts and attaches fields from the account (descriptor, deviceState, symbol) to the tx object
+ *
+ * @param {AccountTransaction} tx
+ * @param {Account} account
+ * @returns {WalletAccountTransaction}
+ */
+export const enhanceTransaction = (
+    tx: AccountTransaction,
+    account: Account,
+): WalletAccountTransaction => {
+    return {
+        descriptor: account.descriptor,
+        deviceState: account.deviceState,
+        symbol: account.symbol,
+        ...tx,
+        // https://bitcoin.stackexchange.com/questions/23061/ripple-ledger-time-format/23065#23065
+        blockTime:
+            account.networkType === 'ripple' && tx.blockTime
+                ? tx.blockTime + 946684800
+                : tx.blockTime,
+        tokens: tx.tokens.map(tok => {
+            return {
+                ...tok,
+                amount: formatAmount(tok.amount, tok.decimals),
+            };
+        }),
+        amount: formatNetworkAmount(tx.amount, account.symbol),
+        fee: formatNetworkAmount(tx.fee, account.symbol),
+        targets: tx.targets.map(tr => {
+            if (typeof tr.amount === 'string') {
+                return {
+                    ...tr,
+                    amount: formatNetworkAmount(tr.amount, account.symbol),
+                };
+            }
+            return tr;
+        }),
+    };
 };
